asdf: Don't symlink lib and bin files
because asdf will only partially work if the ASDF_DIR is resolved to
`/usr/local/`. E.g. just running `asdf` will print

    cat: /usr/local/VERSION: No such file or directory
    version:

    cat: /usr/local/help.txt: No such file or directory

    PLUGIN direnv
      asdf direnv
      asdf direnv hook asdf

    "Late but latest"
    -- Rajinikanth

instead of the version and the complete help message.

It should be installed as the instructions in
https://asdf-vm.com/#/core-manage-asdf?id=add-to-your-shell
as the ASDF_DIR is resolved in
https://github.com/asdf-vm/asdf/blob/ccf76dd7e60de8a3de03cd3574f570dc1f9dbc55/lib/utils.bash#L24-L29
after utils.bash is sourced in
https://github.com/asdf-vm/asdf/blob/ccf76dd7e60de8a3de03cd3574f570dc1f9dbc55/bin/asdf#L4

So to resolve the ASDF_DIR correctly as `/usr/local/opt/asdf`, the
`/usr/local/opt/asdf/bin/asdf` script must be directly executed.

If lib and bin files were symlinked instead, then `/usr/local/bin/asdf`
would be executed, resolving the ASDF_DIR incorrectly to `/usr/local`.